+++ changelog.md
@@ -8,6 +8,9 @@
 - Possibly fixed minor compatibility issue by removing unnecessary use of `AbortSignal.timeout()`
 - Goto will now work on any submit event including enter, but there are
   outstanding problems with focus being annoyingly lost
+- Improve labels for brightness, contrast and saturation
+  - By adding percentage unit (%)
+  - Adding the default value to the tooltip
 
 ## [4.1.4] - 2023-03-13